Moved Queue to a separate module
