+++ vulkano/src/command_buffer/synced/mod.rs
@@ -76,7 +76,8 @@ use super::{
 };
 use crate::{
     buffer::{sys::UnsafeBuffer, BufferAccess},
-    device::{Device, DeviceOwned, Queue},
+    device::{Device, DeviceOwned},
+    queue::Queue,
     image::{sys::UnsafeImage, ImageAccess, ImageLayout, ImageSubresourceRange},
     sync::{
         AccessCheckError, AccessError, AccessFlags, GpuFuture, PipelineMemoryAccess, PipelineStages,
