+++ vulkano/src/sync/future/mod.rs
@@ -22,7 +22,8 @@ use crate::{
         },
         CommandBufferExecError, CommandBufferExecFuture, PrimaryCommandBuffer,
     },
-    device::{DeviceOwned, Queue},
+    device::DeviceOwned,
+    queue::Queue,
     image::{sys::UnsafeImage, ImageLayout},
     swapchain::{self, PresentFuture, PresentRegion, Swapchain},
     DeviceSize, OomError,
