+++ vulkano/src/sync/future/now.rs
@@ -11,7 +11,8 @@ use super::{AccessCheckError, FlushError, GpuFuture};
 use crate::{
     buffer::sys::UnsafeBuffer,
     command_buffer::submit::SubmitAnyBuilder,
-    device::{Device, DeviceOwned, Queue},
+    device::{Device, DeviceOwned},
+    queue::Queue,
     image::{sys::UnsafeImage, ImageLayout},
     sync::{AccessFlags, PipelineStages},
     DeviceSize,
