+++ vulkano/src/sync/mod.rs
@@ -103,7 +103,7 @@
 //! TODO: talk about fence + semaphore simultaneously
 //! TODO: talk about using fences to clean up
 
-use crate::device::Queue;
+use crate::queue::Queue;
 use std::sync::Arc;
 
 pub use self::{
